+++ packages/pangea-sdk/tests/unit/__init__.py
@@ -1,3 +1,4 @@
 from .test_enum import TestEnums
+from .test_file_scan import TestFileScan
 from .test_models import TestModels
 from .test_signer import TestSigner